docker-airflow/requirements.txt
###### Requirements without Version Specifiers ######
# nose
# nose-cov
# beautifulsoup4
apache-airflow
pandas
numpy
google-cloud-storage
google-api-python-client
google-cloud
google-cloud-vision
google.cloud.bigquery

###### Requirements with Version Specifiers ######
# docopt == 0.6.1             # Version Matching. Must be version 0.6.1
# keyring >= 4.1.1            # Minimum version 4.1.1
# coverage != 3.5             # Version Exclusion. Anything except version 3.5
# Mopidy-Dirble ~= 1.1        # Compatible release. Same as >= 1.1, == 1.
